Split parallel loop into their files
